[ADDED] User model, controller, controller test help #10
Also moved default api controller methods to concern.
