--- app/helpers/api/v1/admins_helper.rb
@@ -1,2 +1,0 @@
-module Api::V1::AdminsHelper
-end
