+++ app/models/concerns/user_types.rb
@@ -1,0 +1,6 @@
+module UserTypes
+  extend ActiveSupport::Concern
+
+  TYPE_OPTIONS = ['Admin']
+
+end
